+++ pet/src/main/java/com/veterinaria/dtos/DetalleVentaResponseDTO.java
@@ -1,5 +1,7 @@
 package com.veterinaria.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,6 +13,6 @@ public class DetalleVentaResponseDTO {
     private Long productoId;
     private String nombreProducto;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double subtotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
 }
